+++ Mammoth/Include/TSEUtil.h
@@ -797,7 +797,7 @@ class CSpaceObjectList
 		void DeleteAll (void) { m_List.DeleteAll(); }
 		void FastAdd (CSpaceObject *pObj, int *retiIndex = NULL) { if (retiIndex) *retiIndex = m_List.GetCount(); m_List.Insert(pObj); }
 		void FastAdd (const CSpaceObjectList &List);
-		bool FindObj (const CSpaceObject *pObj, size_t *retiIndex = NULL) const { return m_List.Find(const_cast<CSpaceObject *>(pObj), retiIndex); }
+		bool FindObj (const CSpaceObject *pObj, INT64*retiIndex = NULL) const { return m_List.Find(const_cast<CSpaceObject *>(pObj), retiIndex); }
 		int GetCount (void) const { return m_List.GetCount(); }
 		CSpaceObject *GetObj (int iIndex) const { return m_List[iIndex]; }
 		CSpaceObject *GetRandomObj (void) const { return (m_List.GetCount() == 0 ? NULL : m_List[mathRandomPtr(0, m_List.GetCount() - 1)]); }
